32 bit address space

+++ rustypsx/src/display/gui/actions.rs
@@ -5,6 +5,6 @@ pub enum GuiAction {
     ClearError,
     StepCycles(u32),
     StepFrames(u32),
-    SetBreakpoint(u16),
-    RemoveBreakpoint(u16),
+    SetBreakpoint(u32),
+    RemoveBreakpoint(u32),
 }
